requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2
anthropic>=0.49,<1
openai>=1.30,<2
duckduckgo-search>=7.0,<8
pycryptodome>=3.20,<4
PyNaCl>=1.5,<2
google-genai>=1.0,<2
yt-dlp>=2024.1
playwright>=1.40,<1.58.1
fpdf2>=2.8.6,<3
weasyprint>=63,<69
